requirements.txt
screeninfo
pillow
pygame
openai
pydantic
pydantic-settings
